Extract KSCSwitcher integration into a separate class
